Moved Player Management methods to "Player Manager", simplified Player Class
