+++ player/playerdata.csv
@@ -1,2 +1,3 @@
 doug,BEGINNER
 jeff,ADVANCED
+sarina,BEGINNER
